+++ release_notes.md
@@ -1,5 +1,7 @@
 # dev
 
+- Removed LDAP support, `--auth.source=ldap`, and other LDAP-related
+  CLI flags
 - Removed the Repo Updater app (hidden behind SG_FEATURE_REPOUPDATER flag).
 - Removed `--graphuplink` flag. Data uplink period is fixed to 10 minutes.
 - To configure a different server to forward metrics and events data to,